Change HeaderList to map[string]string
