+++ signature_test.go
@@ -1,63 +1,20 @@
 package httpsignatures
 
-import (
-	"github.com/stretchr/testify/assert"
-	"net/http"
-	"testing"
-)
+// import (
+// 	"github.com/stretchr/testify/assert"
+// 	"net/http"
+// 	"testing"
+// )
 
 const (
 	testSignature = `keyId="Test",algorithm="hmac-sha256",signature="QgoCZTOayhvFBl1QLXmFOZIVMXC0Dujs5ODsYVruDPI="`
-	// testHash      = `JldXnt8W9t643M2Sce10gqCh/+E7QIYLiI+bSjnFBGCti7s+mPPvOjVb72sbd1FjeOUwPTDpKbrQQORrm+xBYfAwCxF3LBSSzORvyJ5nRFCFxfJ3nlQD6Kdxhw8wrVZX5nSem4A/W3C8qH5uhFTRwF4ruRjh+ENHWuovPgO/HGQ=`
-	testHash  = `QgoCZTOayhvFBl1QLXmFOZIVMXC0Dujs5ODsYVruDPI=`
-	testKey   = "U29tZXRoaW5nUmFuZG9t"
-	testDate  = "Thu, 05 Jan 2012 21:31:40 GMT"
-	testKeyID = "Test"
+	testHash      = `QgoCZTOayhvFBl1QLXmFOZIVMXC0Dujs5ODsYVruDPI=`
+	testKey       = "U29tZXRoaW5nUmFuZG9t"
+	testDate      = "Thu, 05 Jan 2012 21:31:40 GMT"
+	testKeyID     = "Test"
 )
 
 // Test
-func TestVerifySignatureFromAuthorizationHeader(t *testing.T) {
-
-	r := &http.Request{
-		Header: http.Header{
-			"Date":              []string{testDate},
-			HeaderAuthorization: []string{authScheme + testSignature},
-		},
-	}
-
-	var v VerificationParameters
-	err := v.FromRequest(r)
-	assert.Nil(t, err)
-	assert.Equal(t, "Test", v.SigParams.KeyID)
-	assert.Equal(t, AlgorithmHmacSha256, v.SigParams.Algorithm)
-	assert.Equal(t, testHash, v.Signature)
-
-	valid, err := v.Verify(testKey, r)
-	assert.Nil(t, err)
-	assert.Equal(t, true, valid)
-}
-
-func TestCreateSignatureFromSignatureHeaderHeader(t *testing.T) {
-	r := http.Request{
-		Header: http.Header{
-			"Date":          []string{testDate},
-			HeaderSignature: []string{testSignature},
-		},
-	}
-
-	var v VerificationParameters
-	err := v.FromRequest(&r)
-	assert.Nil(t, err)
-
-	assert.Equal(t, "Test", v.SigParams.KeyID)
-	assert.Equal(t, AlgorithmHmacSha256, v.SigParams.Algorithm)
-	// assert.Equal(t, testHash, v.Signature)
-
-	valid, err := v.Verify(testKey, &r)
-	assert.Nil(t, err)
-	assert.Equal(t, true, valid)
-}
-
 // func TestCreateSignatureWithNoSignature(t *testing.T) {
 // 	r := http.Request{
 // 		Header: http.Header{
@@ -65,10 +22,12 @@ func TestCreateSignatureFromSignatureHeaderHeader(t *testing.T) {
 // 		},
 // 	}
 
-// 	var s SignatureParameters
-// 	err := s.FromRequest(&r)
+// 	var v VerificationParameters
+// 	err := v.FromRequest(&r)
 // 	assert.Equal(t, ErrorNoSignatureHeader, err)
-// 	assert.Equal(t, SignatureParameters{}, s)
+// 	sigParam := SignatureParameters{Headers: HeaderList{"date": testDate}}
+// 	assert.Equal(t, VerificationParameters{SigParams: &sigParam, Signature: ""}, v)
+// 	assert.Equal(t, SignatureParameters{}, v)
 // }
 
 // func TestValidRequestIsValid(t *testing.T) {
